Refactor the timeline implementation using seperate cells for different parts of timeline.
Summary
In the latest commit, there exists a little delay  when scrolling timeline, to avoid such problem,
change the implementation which no dynamic auto layout constraints update and UIView modification
on runtime, to archive a good uer experience when scrolling.

Implement the commands cell in post timeline,to complete the visual implementation.