Add a fadein and fadeout segue methods.
Summary:
In the storyboard, create a custom type segue to destination view conroller, and set the identifier as "DOPFadeOutSegueForward",
It's all. run the appliation and tap on the photos, the timeline view will fadeout and segue forward to destination viewcontroller.
To segue back, in the destination ViewController, create a unwind connection with the Exit button.